+++ crates/ruma-common/CHANGELOG.md
@@ -9,6 +9,7 @@ Improvements:
 
 - `MatrixVersion` implements `PartialOrd` and `Ord`. The variants are ordered by
   release date, with a newer version being greater than an older version.
+- `Signatures` implements `IntoIterator`
 
 # 0.15.1